requirements.txt
# Core dependencies
PyGithub==2.1.1
python-dotenv==1.0.0

# AI/ML dependencies for embeddings and similarity
openai==1.12.0
numpy==1.26.3
scikit-learn==1.4.0

# Data processing
pandas==2.2.0

# CLI
click==8.1.7

# Testing
pytest==7.4.3
